requirements.txt
numpy~=1.26.4
torchvision~=0.17.1
keras-resnet==0.2.0
tensorflow~=2.15.0